+++ examples/purchaseTesterTypescript/app/RootStackParamList.tsx
@@ -13,6 +13,7 @@ type RootStackParamList = {
     offering: PurchasesOffering | null;
     fontFamily?: string | null;
   };
+  CustomerCenterScreen: { shouldShowCloseButton?: boolean };
 };
 
 export default RootStackParamList;