+++ examples/purchaseTesterTypescript/app/screens/HomeScreen.tsx
@@ -1,6 +1,7 @@
 import React, {useEffect, useState} from 'react';
 
 import {
+  Modal,
   Alert,
   SafeAreaView,
   ScrollView,
@@ -56,6 +57,8 @@ const HomeScreen: React.FC<Props> = ({navigation}) => {
     // from a non-asyn function
     setTimeout(fetchData, 1);
   }, []);
+
+  const [isModalVisible, setIsModalVisible] = useState(false);
 
   // Gets customer info, app user id, if user is anonymous, and offerings
   const fetchData = async () => {
@@ -195,8 +198,31 @@ const HomeScreen: React.FC<Props> = ({navigation}) => {
     }
   };
 
+const onDismissCustomerCenter = () => {
+  setIsModalVisible(false)
+}
+
+const onCustomActionSelected = (event: {actionId: string}) => {
+  console.log('Custom action selected:', event.actionId);
+}
+
   return (
     <SafeAreaView>
+      <Modal
+        animationType="slide"
+        transparent={true}
+        visible={isModalVisible}
+        onRequestClose={onDismissCustomerCenter}
+      >
+        <View style={styles.modalOverlay}>
+          <RevenueCatUI.CustomerCenterView
+            style={styles.modalContent}
+            shouldShowCloseButton={true}
+            onDismiss={onDismissCustomerCenter}
+            onCustomActionSelected={onCustomActionSelected}
+          />
+        </View>
+      </Modal>
       <ScrollView contentInsetAdjustmentBehavior="automatic">
         <View>
           <TouchableOpacity
@@ -269,6 +295,14 @@ const HomeScreen: React.FC<Props> = ({navigation}) => {
           <TouchableOpacity
             onPress={() => navigation.navigate('PaywallModalWithHeader', {})}>
             <Text style={styles.otherActions}>Go to Paywall Modal (with header)</Text>
+          </TouchableOpacity>
+          <TouchableOpacity
+            onPress={() => navigation.navigate('CustomerCenterModalNoHeader', { shouldShowCloseButton: true })}>
+            <Text style={styles.otherActions}>Go to Customer Center Modal (no header)</Text>
+          </TouchableOpacity>
+          <TouchableOpacity
+            onPress={() => navigation.navigate('CustomerCenterModalWithHeader', { shouldShowCloseButton: false })}>
+            <Text style={styles.otherActions}>Go to Customer Center Modal (with header)</Text>
           </TouchableOpacity>
           <TouchableOpacity
             onPress={() =>
@@ -359,6 +393,12 @@ const HomeScreen: React.FC<Props> = ({navigation}) => {
                       } else {
                         console.log('🔍 CUSTOMER CENTER - Management option selected:', option);
                       }
+                    },
+                    onManagementOptionSelected: ({option, actionId, purchaseIdentifier}: CustomerCenterManagementOptionEvent) => {
+                        console.log('🔍 CUSTOMER CENTER - Management option selected:', option, 'with action:', actionId, 'with purchase:', purchaseIdentifier);
+                    },
+                    onCustomActionSelected: ({actionId}: {actionId: string}) => {
+                      console.log('🎯 CUSTOMER CENTER - Custom action selected:', actionId);
                     }
                   }
                 });
@@ -368,6 +408,12 @@ const HomeScreen: React.FC<Props> = ({navigation}) => {
               }
             }}>
             <Text style={styles.otherActions}>Present customer center</Text>
+          </TouchableOpacity>
+          <TouchableOpacity onPress={() => navigation.navigate('CustomerCenterScreen', { shouldShowCloseButton: false })}>
+            <Text style={styles.otherActions}>Go to Customer Center (push)</Text>
+          </TouchableOpacity>
+          <TouchableOpacity onPress={() => setIsModalVisible(true)}>
+            <Text style={styles.otherActions}>Go to Customer Center (modal)</Text>
           </TouchableOpacity>
         </View>
 
@@ -414,6 +460,17 @@ const styles = StyleSheet.create({
     textAlign: 'center',
     color: 'dodgerblue',
     marginVertical: 5,
+  },
+  modalOverlay: {
+    flex: 1,
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  modalContent: {
+    width: '100%',
+    height: '100%',
+    backgroundColor: 'white'
   },
 });