+++ react-native-purchases-ui/android/src/main/java/com/revenuecat/purchases/react/ui/customercenter/events/CustomerCenterEventName.kt
@@ -1,0 +1,6 @@
+package com.revenuecat.purchases.react.ui.customercenter.events
+
+internal enum class CustomerCenterEventName(val eventName: String) {
+    ON_DISMISS("onDismiss"),
+    ON_CUSTOM_ACTION_SELECTED("onCustomActionSelected");
+}