+++ react-native-purchases-ui/ios/CustomerCenterViewWrapper.h
@@ -1,0 +1,22 @@
+//
+//  CustomerCenterViewManager.h
+//
+//  Created by Facundo Menzella on 29/12/23.
+//  Copyright © 2023 Facebook. All rights reserved.
+
+#import <UIKit/UIKit.h>
+#import <React/RCTComponent.h>
+
+@import RevenueCatUI;
+@import PurchasesHybridCommonUI;
+
+@interface CustomerCenterViewWrapper : UIView<RCCustomerCenterViewControllerDelegateWrapper>
+
+@property (nonatomic, copy) RCTDirectEventBlock onDismiss;
+@property (nonatomic, copy) RCTDirectEventBlock onCustomActionSelected;
+@property (nonatomic, assign) BOOL shouldShowCloseButton;
+
+- (instancetype)initWithCustomerCenterViewController:(CustomerCenterUIViewController *)viewController;
+
+@end
+