+++ CHANGELOG.md
@@ -1,5 +1,103 @@
+## 9.3.0
+## RevenueCat SDK
+> [!WARNING]  
+> If you don't have a login system in your app, please make sure your one-time purchase products have been correctly configured in the RevenueCat dashboard as either consumable or non-consumable. If they're incorrectly configured as consumables, RevenueCat will consume these purchases. This means that users won't be able to restore them from version 9.0.0 onward.
+> Non-consumables are products that are meant to be bought only once, for example, lifetime subscriptions.
+
+### ✨ New Features
+* Virtual Currency Support for React Native Web (#1378) via Will Taylor (@fire-at-will)
+### 🐞 Bugfixes
+* Fix `react-native-purchases-ui` preview mode (#1383) via Toni Rico (@tonidero)
+### 📦 Dependency Updates
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 17.3.0 (#1380) via RevenueCat Git Bot (@RCGitBot)
+  * [Android 9.4.0](https://github.com/RevenueCat/purchases-android/releases/tag/9.4.0)
+  * [iOS 5.36.0](https://github.com/RevenueCat/purchases-ios/releases/tag/5.36.0)
+* [RENOVATE] Update dependency com.android.tools.build:gradle to v8.12.2 (#1382) via RevenueCat Git Bot (@RCGitBot)
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 17.2.0 (#1379) via RevenueCat Git Bot (@RCGitBot)
+  * [Android 9.4.0](https://github.com/RevenueCat/purchases-android/releases/tag/9.4.0)
+  * [iOS 5.36.0](https://github.com/RevenueCat/purchases-ios/releases/tag/5.36.0)
+* Updates purchases-hybrid-common to 17.1.0 (#1377) via RevenueCat Git Bot (@RCGitBot)
+  * [Android 9.4.0](https://github.com/RevenueCat/purchases-android/releases/tag/9.4.0)
+  * [iOS 5.36.0](https://github.com/RevenueCat/purchases-ios/releases/tag/5.36.0)
+
+## RevenueCatUI SDK
+### Customer Center
+#### ✨ New Features
+* Added custom actions support for customer center (#1370) via Facundo Menzella (@facumenzella)
+
+### 🔄 Other Changes
+* Add warning with 9.x issues to all versions since 9.0.0 in CHANGELOG (#1385) via Toni Rico (@tonidero)
+* Fix PHC bumping of purchases-js-hybrid-mappings (#1375) via Toni Rico (@tonidero)
+
+## 9.2.3
+## RevenueCat SDK
+> [!WARNING]  
+> If you don't have any login system in your app, please make sure your one-time purchase products have been correctly configured in the RevenueCat dashboard as either consumable or non-consumable. If they're incorrectly configured as consumables, RevenueCat will consume these purchases. This means that users won't be able to restore them from version 9.0.0 onward.
+> Non-consumables are products that are meant to be bought only once, for example, lifetime subscriptions.
+
+### 📦 Dependency Updates
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 17.0.0 (#1372) via RevenueCat Git Bot (@RCGitBot)
+  * [Android 9.3.2](https://github.com/RevenueCat/purchases-android/releases/tag/9.3.2)
+  * [Android 9.3.1](https://github.com/RevenueCat/purchases-android/releases/tag/9.3.1)
+  * [Android 9.3.0](https://github.com/RevenueCat/purchases-android/releases/tag/9.3.0)
+  * [iOS 5.35.1](https://github.com/RevenueCat/purchases-ios/releases/tag/5.35.1)
+  * [iOS 5.35.0](https://github.com/RevenueCat/purchases-ios/releases/tag/5.35.0)
+* [RENOVATE] Update dependency com.android.tools.build:gradle to v8.12.1 (#1367) via RevenueCat Git Bot (@RCGitBot)
+
+### 🔄 Other Changes
+* Bump tmp from 0.2.3 to 0.2.4 (#1355) via dependabot[bot] (@dependabot[bot])
+
+## 9.2.2
+## RevenueCat SDK
+> [!WARNING]  
+> If you don't have any login system in your app, please make sure your one-time purchase products have been correctly configured in the RevenueCat dashboard as either consumable or non-consumable. If they're incorrectly configured as consumables, RevenueCat will consume these purchases. This means that users won't be able to restore them from version 9.0.0 onward.
+> Non-consumables are products that are meant to be bought only once, for example, lifetime subscriptions.
+
+### 🐞 Bugfixes
+* Remove `react-native-get-random-values` dependency temporarily (#1364) via Toni Rico (@tonidero)
+
+## 9.2.1
+## RevenueCat SDK
+> [!WARNING]  
+> If you don't have any login system in your app, please make sure your one-time purchase products have been correctly configured in the RevenueCat dashboard as either consumable or non-consumable. If they're incorrectly configured as consumables, RevenueCat will consume these purchases. This means that users won't be able to restore them from version 9.0.0 onward.
+> Non-consumables are products that are meant to be bought only once, for example, lifetime subscriptions.
+
+### 📦 Dependency Updates
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 16.2.2 (#1361) via RevenueCat Git Bot (@RCGitBot)
+### Customer Center
+#### 🐞 Bugfixes
+* fix: Support modal navigation to CustomerCenter (#1209) via Facundo Menzella (@facumenzella)
+
+### 🔄 Other Changes
+* Update MagicWeather to RN 0.81 (#1357) via Jens-Fabian Goetzmann (@jefago)
+
+## 9.2.0
+## RevenueCat SDK
+> [!WARNING]  
+> If you don't have any login system in your app, please make sure your one-time purchase products have been correctly configured in the RevenueCat dashboard as either consumable or non-consumable. If they're incorrectly configured as consumables, RevenueCat will consume these purchases. This means that users won't be able to restore them from version 9.0.0 onward.
+> Non-consumables are products that are meant to be bought only once, for example, lifetime subscriptions.
+
+### ✨ New Features
+* Add browser support through Web Billing (purchases-js) (#1330) via Toni Rico (@tonidero)
+### 📦 Dependency Updates
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 16.2.0 (#1352) via RevenueCat Git Bot (@RCGitBot)
+  * [Android 9.2.0](https://github.com/RevenueCat/purchases-android/releases/tag/9.2.0)
+  * [iOS 5.34.0](https://github.com/RevenueCat/purchases-ios/releases/tag/5.34.0)
+* [RENOVATE] Update dependency com.android.tools.build:gradle to v8.12.0 (#1349) via RevenueCat Git Bot (@RCGitBot)
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 16.1.0 (#1346) via RevenueCat Git Bot (@RCGitBot)
+  * [Android 9.2.0](https://github.com/RevenueCat/purchases-android/releases/tag/9.2.0)
+  * [iOS 5.34.0](https://github.com/RevenueCat/purchases-ios/releases/tag/5.34.0)
+
+### 🔄 Other Changes
+* Add warning on CHANGELOG (#1351) via Toni Rico (@tonidero)
+* Improve 9.0.0 changelog (#1347) via Toni Rico (@tonidero)
+
 ## 9.1.0
 ## RevenueCat SDK
+> [!WARNING]  
+> If you don't have any login system in your app, please make sure your one-time purchase products have been correctly configured in the RevenueCat dashboard as either consumable or non-consumable. If they're incorrectly configured as consumables, RevenueCat will consume these purchases. This means that users won't be able to restore them from version 9.0.0 onward.
+> Non-consumables are products that are meant to be bought only once, for example, lifetime subscriptions.
+
 ### ✨ New Features
 * Support Virtual Currencies (#1336) via Will Taylor (@fire-at-will)
 ### 📦 Dependency Updates